prevent app crashes
