--- app/report-data/get-AR-report-data.js
@@ -1,24 +1,0 @@
-const db = require('../data')
-
-const getARReportData = async (startDate, endDate) => {
-  const whereClause = {
-    arValue: {
-      [db.Sequelize.Op.ne]: null
-    }
-  }
-
-  if (startDate && endDate) {
-    whereClause.lastUpdated = {
-      [db.Sequelize.Op.between]: [startDate, endDate]
-    }
-  }
-
-  return db.reportData.findAll({
-    where: whereClause,
-    raw: true
-  })
-}
-
-module.exports = {
-  getARReportData
-}
