+++ app/server/routes/ap-report-data.js
@@ -1,5 +1,6 @@
+const { AP } = require('../../constants/ledgers')
 const { GET } = require('../../constants/methods')
-const { getAPReportData } = require('../../report-data/get-AP-report-data')
+const { getAPARReportData } = require('../../report-data/get-AP-AR-report-data')
 
 module.exports = {
   method: GET,
@@ -8,7 +9,7 @@ module.exports = {
     handler: async (request, h) => {
       const startDate = request.query.startDate ? new Date(request.query.startDate) : null
       const endDate = request.query.endDate ? new Date(request.query.endDate) : null
-      const apReportData = await getAPReportData(startDate, endDate)
+      const apReportData = await getAPARReportData(startDate, endDate, AP)
       return h.response({ apReportData })
     }
   }
