+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ffc-pay-tracking",
-  "version": "1.2.4",
+  "version": "1.2.5",
   "description": "Track payments and publish notifications",
   "homepage": "https://github.com/DEFRA/ffc-pay-tracking",
   "main": "app/index.js",
