+++ test/unit/app/server/routes/ap-report-data.test.js
@@ -1,12 +1,12 @@
-const { getAPReportData } = require('../../../../../app/report-data/get-AP-report-data')
+const { getAPARReportData } = require('../../../../../app/report-data/get-AP-AR-report-data')
 const { options } = require('../../../../../app/server/routes/ap-report-data')
 
-jest.mock('../../../../../app/report-data/get-AP-report-data')
+jest.mock('../../../../../app/report-data/get-AP-AR-report-data')
 
 describe('GET /ap-report-data', () => {
   test('should return AP report data when startDate and endDate are provided', async () => {
     const mockData = [{ apValue: 1 }, { apValue: 2 }]
-    getAPReportData.mockResolvedValue(mockData)
+    getAPARReportData.mockResolvedValue(mockData)
 
     const startDate = '2022-01-01'
     const endDate = '2022-12-31'
@@ -27,7 +27,7 @@ describe('GET /ap-report-data', () => {
 
   test('should return AP report data when startDate and endDate are not provided', async () => {
     const mockData = [{ apValue: 1 }, { apValue: 2 }]
-    getAPReportData.mockResolvedValue(mockData)
+    getAPARReportData.mockResolvedValue(mockData)
 
     const mockRequest = {
       query: {}
