+++ test/unit/app/server/routes/ar-report-data.test.js
@@ -1,12 +1,12 @@
-const { getARReportData } = require('../../../../../app/report-data/get-AR-report-data')
+const { getAPARReportData } = require('../../../../../app/report-data/get-AP-AR-report-data')
 const { options } = require('../../../../../app/server/routes/ar-report-data')
 
-jest.mock('../../../../../app/report-data/get-AR-report-data')
+jest.mock('../../../../../app/report-data/get-AP-AR-report-data')
 
 describe('GET /ar-report-data', () => {
   test('should return AR report data when startDate and endDate are provided', async () => {
     const mockData = [{ arValue: 1 }, { arValue: 2 }]
-    getARReportData.mockResolvedValue(mockData)
+    getAPARReportData.mockResolvedValue(mockData)
 
     const startDate = '2022-01-01'
     const endDate = '2022-12-31'
@@ -27,7 +27,7 @@ describe('GET /ar-report-data', () => {
 
   test('should return AR report data when startDate and endDate are not provided', async () => {
     const mockData = [{ arValue: 1 }, { arValue: 2 }]
-    getARReportData.mockResolvedValue(mockData)
+    getAPARReportData.mockResolvedValue(mockData)
 
     const mockRequest = {
       query: {}
